+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # Changelog
 
-## [Unreleased] 2024-07-23
+## [Unreleased] 2024-07-29
 
+- Brought back the Vertex AI Pipelines scheduling capability
 - Migrated to kfp 2
 - Removed `image_pull_policy` parameter from configuration, as it only applies to Kubernetes backend and not Vertex AI,
 and it's only available in `kfp-kubernetes` extension package